update:conversions

+++ exercises/conversions/using_as.rs
@@ -11,10 +11,9 @@
 // hint.
 
 // I AM NOT DONE
-
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
